Update Dockerfile and compose-dev.yaml
This commit updates the Dockerfile to include the necessary permissions for the Plugin directory and copies the custom WordPress plugin to the appropriate location in the container. It also updates the compose-dev.yaml file to use the latest version of the WordPress image and correctly mount the init-script.sh file.